requirements.txt
numpy==1.24.2
opencv-python==4.7.0.72
Pillow==9.5.0
flask==2.2.3
flask-sqlalchemy==3.0.3
scipy==1.10.1